Updates for Issue #6 and other WIP
Added the debugging.txt file which details how debugging works in SpecBAS.

Added SpecBAS reference document (outdated, needs to be updated soon).

Fixed some overflow errors.

Fixed Issue #6 - crash when loading sprites due to an overflow when copying frame data to the sprite bank.

Fixed an issue with recent files not showing if you load your startupsequence file.

Added conditional and data breakpoints

Added more functionality to breakpoint editor (ctrl+n) but actual BP management is still WIP.

Placed all debugging procedures behins a "lock" to minimise performance impact when not debugging.

Fixed status label in the editor *again*

Fixed mouse not moving when SCREEN LOCK is in use

Much faster startup, no more black window.

+++ src/MainForm.pas
@@ -198,8 +198,8 @@ Begin
     GetCursorPos(p);
     p := Main.ScreenToClient(p);
     {$IFDEF OpenGL}
-      MOUSEX := Round(p.X / ScaleMouseX);
-      MOUSEY := Round(p.Y / ScaleMouseY);
+      MOUSEX := Integer(Round(p.X / ScaleMouseX));
+      MOUSEY := Integer(Round(p.Y / ScaleMouseY));
     {$ELSE}
       MOUSEX := p.X;
       MOUSEY := p.Y;
